// src/types/Items/ItemsCardProps.tsx
export default interface ItemsCardProps {
  title: string;
  description: string;
  img: string;
  manufacture: string;
  brand: string;
  index: string;
  cover: string;
  color: string;
  diameter: number;
  material: string;
  geometry: string;
  lensType: string;
  lensClass: string;
  technology: string;
  dioptre: number;
  cylinder: number;
  add: number;
  price: number;
  amount: number;
  department: number;
}